Fix faulty test.
